Added write ahead log

+++ bboxdb-server/src/main/java/org/bboxdb/storage/wal/WalManager.java
@@ -32,7 +32,7 @@ public class WalManager {
 	 * @param memtableNumber
 	 * @return
 	 */
-	public static File getFileForWal(final File basedir, final int memtableNumber) {
+	public static File getFileForWal(final File basedir, final long memtableNumber) {
 		return new File(basedir.getAbsolutePath() + "/" + "wal_" 
 				+ memtableNumber + SSTableConst.MEMTABLE_WAL_SUFFIX);
 	}
